requirements.txt
streamlit
pandas
numpy
mysql.connector.python